fix(SDK): Add IIFE to the `dot-uve` file (#31896)
## Notes

I know there is a way to tell `esbuild` to use the `iife` format, but NX
has an [issue](https://github.com/nrwl/nx/discussions/23427) where all
the NX config is getting precedence over the `esbuild` config, causing
the format to always fall into `esm` or `cjs`, that's why I used the
`banner` and `footer` approach which is not overridden by NX.

## Summary

This pull request includes changes to the
`core-web/libs/sdk/uve/project.json` and
`dotCMS/src/main/webapp/ext/uve/dot-uve.js` files to enhance the
functionality and structure of the code. The most important changes
include adding esbuild options in the project configuration and wrapping
the existing JavaScript code in an immediately invoked function
expression (IIFE).

Enhancements to project configuration:

*
[`core-web/libs/sdk/uve/project.json`](diffhunk://#diff-cada4c4405fb3f2ee3b60de34daab60df5a4dfffda1f6ef66bd6bbae72df4beaL59-R67):
Added `esbuildOptions` with `banner` and `footer` properties to wrap the
JavaScript code in an IIFE.

Code structure improvements:

*
[`dotCMS/src/main/webapp/ext/uve/dot-uve.js`](diffhunk://#diff-c3400a484a715c8da9dae0ee7898dad9a10ddc5c75b4b6043f3499a3388e431fL1-R3):
Wrapped the entire JavaScript code in an IIFE to avoid polluting the
global scope.

## Screenshot


https://github.com/user-attachments/assets/8c8e37f7-9859-4a5c-ac9b-a5e354d59869