chore(uve): improves in UVE to work with dot-uve.js file (#31902)
This PR introduces an improvement to UVE that allows applications using
only dot-uve.js to properly display pages in LIVE mode.
Previously, the loading screen wouldn't disappear, preventing the page
from rendering.

Additionally, it fixes a bug where a modal would incorrectly open when
navigating to traditional pages.




https://github.com/user-attachments/assets/ca1419a2-944b-4945-94ee-6071ce1d63a7




This pull request includes several changes to enhance the functionality
and error handling of the Universal Visual Editor (UVE) SDK in the
DotCMS platform. The most important changes include the addition of the
`UVE_MODE` import, modifications to the `EditEmaEditorComponent` and
`withEditor` tests to include the `mode` parameter, and improvements to
error handling in the `withLoad` function.

### Enhancements to UVE SDK:

*
[`core-web/libs/portlets/edit-ema/portlet/src/lib/edit-ema-editor/edit-ema-editor.component.spec.ts`](diffhunk://#diff-34ddc5fbacaf04b962f2037385ed284310d5faf35ba409d5705b2caadd5d796aR80):
Added `UVE_MODE` import and updated test cases to include the `mode`
parameter.
[[1]](diffhunk://#diff-34ddc5fbacaf04b962f2037385ed284310d5faf35ba409d5705b2caadd5d796aR80)
[[2]](diffhunk://#diff-34ddc5fbacaf04b962f2037385ed284310d5faf35ba409d5705b2caadd5d796aR409)
[[3]](diffhunk://#diff-34ddc5fbacaf04b962f2037385ed284310d5faf35ba409d5705b2caadd5d796aL2627-R2629)

*
[`core-web/libs/portlets/edit-ema/portlet/src/lib/store/features/editor/withEditor.spec.ts`](diffhunk://#diff-48915538475425e5656151bb2a73df7a99b4c80c2817c085f96352ef3252d5cbL45-R46):
Updated initial state and test cases to include the `mode` parameter.
[[1]](diffhunk://#diff-48915538475425e5656151bb2a73df7a99b4c80c2817c085f96352ef3252d5cbL45-R46)
[[2]](diffhunk://#diff-48915538475425e5656151bb2a73df7a99b4c80c2817c085f96352ef3252d5cbL105-R106)
[[3]](diffhunk://#diff-48915538475425e5656151bb2a73df7a99b4c80c2817c085f96352ef3252d5cbL217-R218)

### Improvements to error handling:

*
[`core-web/libs/portlets/edit-ema/portlet/src/lib/store/features/load/withLoad.ts`](diffhunk://#diff-0cbc04b5047422551dde0970e4ea1d0e1fcce3d9d8f4e0b4612fbc30b1658487L104-R107):
Enhanced error handling by logging errors and extracting the error
status in the `withLoad` function.
[[1]](diffhunk://#diff-0cbc04b5047422551dde0970e4ea1d0e1fcce3d9d8f4e0b4612fbc30b1658487L104-R107)
[[2]](diffhunk://#diff-0cbc04b5047422551dde0970e4ea1d0e1fcce3d9d8f4e0b4612fbc30b1658487L124-R130)
[[3]](diffhunk://#diff-0cbc04b5047422551dde0970e4ea1d0e1fcce3d9d8f4e0b4612fbc30b1658487L193-R202)

### Code refactoring:

*
[`core-web/libs/sdk/uve/src/script/sdk-editor.ts`](diffhunk://#diff-e51b90694192753f2ecf8982c7223da00a4a2ed8564a029289a1f8cbd60e2df7L9-R16):
Refactored imports to include specific functions from `public` and
updated the global `dotUVE` object.
[[1]](diffhunk://#diff-e51b90694192753f2ecf8982c7223da00a4a2ed8564a029289a1f8cbd60e2df7L9-R16)
[[2]](diffhunk://#diff-e51b90694192753f2ecf8982c7223da00a4a2ed8564a029289a1f8cbd60e2df7L20-R32)

*
[`dotCMS/src/main/webapp/ext/uve/dot-uve.js`](diffhunk://#diff-c3400a484a715c8da9dae0ee7898dad9a10ddc5c75b4b6043f3499a3388e431fR1-R12):
Added a new header comment for the UVE SDK and refactored the existing
code for better readability and maintainability.

---------

Co-authored-by: Kevin Davila <56242609+kevindaviladev@users.noreply.github.com>