fix(UVE): Add Params to Router Link to preserve state across portlets (#31923)
This pull request enhances the functionality of the
`EditEmaNavigationBarComponent` by adding support for query parameters
and improving its test coverage. The most significant changes include
updates to the component's HTML template, the addition of query
parameter handling in the TypeScript file, and new test cases to
validate the functionality.

### Enhancements to query parameter handling:

* **HTML Template (`edit-ema-navigation-bar.component.html`)**: Added
support for query parameters using the `$params()` method and configured
`routerLinkActiveOptions` to ignore query parameters, matrix parameters,
and fragments while matching paths exactly.
* **Component Logic (`edit-ema-navigation-bar.component.ts`)**:
Introduced a new `$params` property in the component to access query
parameters from the store.

### Test coverage improvements:

* **Mock Data (`edit-ema-navigation-bar.component.spec.ts`)**: Added a
new `pageParams` method to the mock store to provide test data for query
parameters.
* **New Test Case (`edit-ema-navigation-bar.component.spec.ts`)**:
Implemented a test to verify that the correct query parameters are
applied when a navigation item is clicked. This ensures the router URL
contains the expected query parameters.


## Screenshot 

https://github.com/user-attachments/assets/f9be757f-fdab-4ee0-a32e-d3977ad00cfe