fix(template-builder): Ignore `gridstack` files on Jest Transform step (#31927)
This pull request includes a small update to the Jest configuration in
the `core-web/libs/template-builder/jest.config.ts` file. The change
modifies the `transformIgnorePatterns` property to include an exception
for the `gridstack` library.

-
[`core-web/libs/template-builder/jest.config.ts`](diffhunk://#diff-ffe331fbd44eff4ad040083f0f2401ec15592608a175d47d85d54e2191f539a7L31-R31):
Updated `transformIgnorePatterns` to
`['node_modules/(?!.*\.mjs$|gridstack)']` to ensure the `gridstack`
library is not ignored during transformation.