fix(import): Fix issue when importing contents with relationship due to language (#31962)
Problem
During the import process, an error was occurring when a contentlet was
related to another contentlet that had versions in multiple languages.
The issue was that the code treats a content with different languages as
distinct entities — it incorrectly assumed they were different contents,
although the content is the same.

Fix
When validating if the content is trying to relate to more contents that
the one already has, we group by identifier the array of the contents
related, this way although the content has more than one language
versions, it would be interpreted as an unique relation. So in the
scenario the issue shows it won't throw an error. If we have a case in
where the content has more than one relation with the same language,
then we are going to get the error.

---------

Co-authored-by: Jose Castro <jose.castro@dotcms.com>