fix(Field API): add validation on CleanUpFieldReferencesJob when saving field (#31983)
If we delete a field and want to create another one with the same var
name we should wait until the CleanUpFieldReferencesJob ends.

So to achieve this, a new validation method `validateFieldReferences`
was added to the save field method, this validation function checks if
there is a CleanUpFieldReferencesJob running with a field with the same
name, if it finds one, then it throws an exception the user will see.

Also, System Messages were added to notify when the job starts and
finishes running.



https://github.com/user-attachments/assets/0379c3b0-145a-45cd-bd00-ed01ebf32a08

---------

Co-authored-by: erickgonzalez <erick.gonzalez@dotcms.com>