fix(deployment): Bumping Docker Buildx version to fix GHA error. (#32057)
## Proposed Changes

This pull request updates the Docker Buildx version across multiple
GitHub Actions workflows to ensure compatibility with the latest
features and improvements. The version is updated from `v0.20.0` to
`v0.23.0`.

Updates to Docker Buildx version in workflows:

*
[`.github/workflows/cicd_manual_build-docker-context.yml`](diffhunk://#diff-13bfcd88b6c32e26f198b20ece9e8a99e917c1e8330e2a5dd08bd06062745301L44-R44):
Updated the `buildx` version to `v0.23.0` in the "Docker Setup Buildx"
step.
*
[`.github/workflows/cicd_manual_build-java-base.yml`](diffhunk://#diff-f9bc49055186ce5607d3cac95f9970a5b48533ccdece0fae8ac5d5670d25c62cL41-R41):
Updated the `buildx` version to `v0.23.0` in the "Docker Setup Buildx"
step.
*
[`.github/workflows/legacy-release_comp_maven-build-docker-image.yml`](diffhunk://#diff-e53524b18be4b475c05830ca3ca5d8da700ad69571f4a563794711f41e466af8L269-R269):
Updated the `buildx` version to `v0.23.0` in the "Docker Setup Buildx"
step.