Fixed color on roles selector (#32128)
https://github.com/user-attachments/assets/e7a9dbd9-5388-4554-8d5d-2b3ae8db607e


This pull request introduces a small visual enhancement to the user
interface in the `view_users.jsp` file. Specifically, it adds a style
rule to improve the appearance of selected options in the
`#userRolesSelect` dropdown.

*
[`dotCMS/src/main/webapp/html/portlet/ext/useradmin/view_users.jsp`](diffhunk://#diff-1b673eafaf482579d77a457632e9f8dd599e45ec32b65715555dee62ac5a6fb3R54-R57):
Added a CSS rule to apply a light-dark background color to selected
options in the `#userRolesSelect` dropdown.

Co-authored-by: Kevin Davila <56242609+kevindaviladev@users.noreply.github.com>