fix(static push): Add implementation to bundle folders (#32210)
**Problem**
The generation of folder tree when static push publishing was not being
considered on the generation of the bundle, so if the folder didn't have
any files to push publish then the bundle generated was empty.

**Fix**
A new `StaticFolderBundler` class was created, this class implements the
generation of the folder tree according to the selected folder to push,
this way now even if the folder is empty or has files that are not
pages, the bundle will be generated.