fix(UVE): SEO and Accessability tools URL (#32283)
This pull request adds test cases to improve coverage for the `UVEStore`
and refines the logic for handling URLs in the `currentUrl` computation.
The changes ensure better handling of edge cases and improve code
reliability.

### Test Coverage Improvements:

* Added new test cases to verify error handling for specific HTTP status
codes (`404`, `403`, and `401`) in the `UVEStore`
(`dot-uve.store.spec.ts`).
[[1]](diffhunk://#diff-900c5f83c0b1dea02a19ae54db7857ee3d11ce7ed9b2df4bf19000b9fe0d6f2bR198)
[[2]](diffhunk://#diff-900c5f83c0b1dea02a19ae54db7857ee3d11ce7ed9b2df4bf19000b9fe0d6f2bR210)
[[3]](diffhunk://#diff-900c5f83c0b1dea02a19ae54db7857ee3d11ce7ed9b2df4bf19000b9fe0d6f2bR222)
* Introduced test cases to validate the behavior of `currentUrl` for
URLs with and without an initial slash (`dot-uve.store.spec.ts`).

### URL Handling Refinements:

* Updated the `currentUrl` computation to add an initial slash only when
the URL does not already have one (`dot-uve.store.ts`).

### Video


https://github.com/user-attachments/assets/4981a17e-66b3-4594-8141-e3e3ad6494cc