fix(templates) : Add containers included by dotParse directive to template (#32295)
Closes #31790

### Proposed Changes
This pull request introduces changes to the directive execution in
Velocity in order to get all the containers included in a template body,
included containers in a separate file processed with the dotParse
directive. These are the specific changes:
* A new flag `_dotDontUseDirectiveCache` is used to disable caching
while getting the containers on the template.
* The `shouldLoadAndRenderTemplate` method has been added to the
`DotDirective` class to allow subclasses to disable template files from
being loaded.
* Also a new flag `_dotDontLoadContainers` is used to disable containers
from being loaded because we are only interested in getting the
container identifiers when checking the template. The `ParseContainer`
class overwrites the `shouldLoadAndRenderTemplate` method to implement
this behaviour.
* The `afterRender` method has been updated in the `ParseCotainer` class
to collect container IDs in the `_dotTemplateContainerIds` variable.
* The `getContainerUUIDFromHTML` method from the `TemplateFactoryImpl`
class has been refactored to parse the Velocity code and to use Velocity
context for extracting container UUIDs, replacing the previous
line-by-line parsing approach.

### Minor UI fixes:
* **ContentSelector.js**: Fixed an issue where the `selectButton` click
event did not stop propagation, ensuring proper event handling. This
stop propagation was removed by the fix for the issue #31601, but the
problem is that causes the event being handled by both the button click
handler and by the row click handler. With this new fix, the appropriate
`scope` variable is used to check the value of
`useRelateContentOnSelect` flag in the `selected` function to check if
the dialog is being used to relate content.

### Checklist
- [x] Tests