fix(core): Adding back support of Bring Back Version feature in Container's portlet (#32205)
### Proposed Changes
* The Bring Back version feature was removed from containers after a
refactor made [here](https://github.com/dotCMS/core/issues/19087). These
changes add it back. The functionality was removed in 2021.
* An additional fix was included in the `ContainerResource.update()`.
The logic was making changes in the object in the cache instead of using
a new container instance, which was messing up the cache
* The `bringBackTemplateVersion` method was renamed to
`bringBackVersion` as it is used by templates and containers too.

### Video



https://github.com/user-attachments/assets/1df72de1-af65-432c-bedc-37ab18e4a22d



Refs: #32127

---------

Co-authored-by: Jalinson Diaz <zjaaaldev@gmail.com>
Co-authored-by: Rafael Velazco <rjvelazco21@gmail.com>
Co-authored-by: Kevin <kfariid@gmail.com>