fix(uve): Support Containers added to Themes (#32389)
This pull request introduces enhancements to the container management
logic in the `edit-ema` portlet, addressing edge cases where containers
are hardcoded in themes but not indexed in the `PageAPI`. Additionally,
it improves type definitions and adds comprehensive test cases to
validate the new behavior.

### Enhancements to container management logic:
* Updated `insertContentletInContainer`,
`deleteContentletFromContainer`, and
`insertPositionedContentletInContainer` functions to handle cases where
containers are hardcoded in themes but not indexed in the `PageAPI`.
These functions now check if the container exists in `pageContainers`
and add it if missing.
[[1]](diffhunk://#diff-2fd4d225bb33327848da73766ec5efbe037e3aed1a1e51a50c19f6a2be8cdb80R75-R89)
[[2]](diffhunk://#diff-2fd4d225bb33327848da73766ec5efbe037e3aed1a1e51a50c19f6a2be8cdb80R125-R139)
[[3]](diffhunk://#diff-2fd4d225bb33327848da73766ec5efbe037e3aed1a1e51a50c19f6a2be8cdb80R203-R217)

### Type definition improvements:
* Added optional properties `acceptTypes`, `maxContentlets`, and
`variantId` to the `PageContainer` interface in `shared/models.ts`.
These properties provide additional configuration options for
containers.

### Test coverage improvements:
* Added test cases to validate the behavior of
`deleteContentletFromContainer`, `insertContentletInContainer`, and
`insertPositionedContentletInContainer` when containers are missing from
`pageContainers`. These tests ensure that containers are correctly added
and contentlets are managed as expected.
[[1]](diffhunk://#diff-6b95047786d13d69ebdcb642c6d382bc35784a2ada53f95d1908119ab08dfb91R157-R201)
[[2]](diffhunk://#diff-6b95047786d13d69ebdcb642c6d382bc35784a2ada53f95d1908119ab08dfb91R354-R450)