chore(Block Editor): Update tiptap library from beta to stable (2.14.0) (#32402)
This pull request updates the dependencies in `core-web/package.json`,
focusing on upgrading the `@tiptap` ecosystem to the latest stable
versions and introducing a new library, `ngx-tiptap`. Additionally, it
removes older `prosemirror` dependencies that are no longer required.

### Dependency Updates:

* Upgraded all `@tiptap` packages from beta versions (`^2.0.0-beta.218`
and `^2.0.0-beta.220`) to stable version `^2.14.0`. This includes core,
extensions, and starter kit packages.
* Updated `@tiptap/pm` to version `^2.14.0` for compatibility with the
latest `@tiptap` packages.

### New Dependency:

* Added `ngx-tiptap` version `^12.0.0`, which provides Angular bindings
for the `@tiptap` editor.

### Removed Dependencies:

* Removed `prosemirror-model` and `prosemirror-view`, as they are no
longer needed with the updated `@tiptap` ecosystem.


### Video


https://github.com/user-attachments/assets/51037c21-1d40-4835-9baf-9c415695584c



https://github.com/user-attachments/assets/a79f29d6-7e5c-4985-807a-473a82923215