fix(graphql) map attribute failure for contentlets within page variants in graphql #32384  (#32427)
### Proposed Changes
* Added a new private method `getHydratedMapWithVariantFallback` in
`ContentMapDataFetcher` to handle content hydration with explicit or
implicit variant resolution. The method includes fallback logic to the
`DEFAULT` variant when a requested variant is unavailable.
* Replaced the direct usage of `DotContentletTransformer` with the new
`getHydratedMapWithVariantFallback` method in the `get` method,
improving content hydration flexibility.

### Checklist
- [x] Tests
- [x] Translations
- [x] Security Implications Contemplated  

### Additional Info
This improves the correctness and maintainability of variant resolution
logic in `_map` queries and ensures coverage for edge cases in
contentlet hydration across variants.

### Screenshots

![image](https://github.com/user-attachments/assets/e87c5418-468c-46df-91a2-da416c5b8c7a)