fix(Apps): add error detail when exporting app config  (#32407)
**Problem**

A generic message was displayed when getting an error in the Apps export
config tab. Also, it wasn't considering the case when the API that was
requesting returned an error.

**Fix**

Added a check that validates if the response has an `error-message` as
headers (this means that the API is returning an error) and if its has,
then throw an error so it can be received correctly, and when received,
display it.

<img width="417" alt="Captura de pantalla 2025-06-12 a la(s) 12 04 01 p
 m"
src="https://github.com/user-attachments/assets/e1ac14c4-a98f-44b2-adaf-91214e4c2f9c"
/>