Issue 32377 analytics endpoint validation (#32486)
The idea is created json files that allow to set the validation that we
want to apply for each event type

for now just the pageview.json validation files was created.
also the all.json files was created to handle general validation rules
that apply for all the event type.

If any validation in the all.json fail then the payload is rejected
completly without even check the events section

### Proposed Changes
* Cretae Analytics Validators


https://github.com/dotCMS/core/pull/32486/files#diff-c76f73bcb7d41f9d8362c06f9d41efd0cce07909751d2b69578324a0d8096234R79-R84

* create util class to load fields validation rules from json files


https://github.com/dotCMS/core/pull/32486/files#diff-c76f73bcb7d41f9d8362c06f9d41efd0cce07909751d2b69578324a0d8096234R69

* create json files with the validation rules

* Create class to validate the payload


https://github.com/dotCMS/core/pull/32486/files#diff-c14a868f5239076820535773e4603ab886255156942ac9b3af648b8b51db0decR16

* class to keep validation rules for each EventType


https://github.com/dotCMS/core/pull/32486/files#diff-c14a868f5239076820535773e4603ab886255156942ac9b3af648b8b51db0decR16

* Rename the languageID attribute to labguage_id


https://github.com/dotCMS/core/pull/32486/files#diff-875d75f7bb3357903411de7c75fc9c447989b8100034fdd1028effff0c9fa177R127