chore(uve): Fixed save contentlets from palette content-types (#32514)
https://github.com/user-attachments/assets/28e4e6d8-48ab-4e4b-85ac-7d48c68640c6

Issue: Previously, we used a conditional check that, due to how it was
written, didn’t properly handle the expected case:

`if (!actionPayload || graphql)
`

In this condition, the evaluation wasn’t based on a strict boolean
check, but rather on the truthiness of the graphql object. As a result,
in headless mode—where graphql always had a value—the condition would
always evaluate as true.

This caused the current content to always be "reloaded," and the request
to save the new contentlet was never actually made.

Now, a preliminary evaluation is performed without needing to combine
actionPayload and graphql in a single condition. All tests are still
passing, and after manual testing, no issues were found as a result of
this change.

---------

Co-authored-by: Kevin <kfariid@gmail.com>