fix(Users Portlet): fix data returned when saving User Details (#32465)
**Problem**

The issue was not limited to users with the Admin role—it occurs when a
user attempts to update their own information. The backend included a
validation that checked if the user requesting the details was the same
as the logged-in user. In such cases, it returned the currently
logged-in user’s data. However, since the logged-in user's data is not
refreshed until logout, the response contained outdated information.

**Fix**

Replaced the return value for the user that was retrieved from the
database instead of the one logged in.