Fix(block-editor) : Fixed removedAtttribute method on velocity request wrapper (#32542)
Closes #32522

### Proposed Changes
* This pull request updates the `removeAttribute` method from the
`VelocityRequestWrapper` class to prevent the removal of attributes
listed in the `SET_VALUE_BLACKLIST`. If an attribute is blacklisted, the
method now exits without invoking the superclass method, otherwise the
attribute is removed.
* Added a new test suite in `VelocityRequestWrapperTest` to verify the
behavior of `setAttribute` and `removeAttribute` methods: ensures
blacklisted attributes cannot be set or removed and confirms that
non-blacklisted attributes can be set and removed as expected.
* With the attribute removal working as expected, the
`refreshBlockEditorDataMap` method from the `StoryBlockAPIImpl`
will work as expected in these lines:

https://github.com/dotCMS/core/blob/484527afeea886a39d45473227d3340a7427dd20/dotCMS/src/main/java/com/dotcms/contenttype/business/StoryBlockAPIImpl.java#L451-L453
That attribute is removed to signal that there are no recursion levels
when refreshing the data map for a contentlet referenced from a block
editor field. When it wasn't removed, content referenced from the second
block editor field wasn't refreshed as expected.

### Checklist
- [x] Tests